# docs/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Provider Agent Liquidity & Coordination Platform",
    "description": "Decision-support prototype API. Phase 2 foundation \u2014 business features are stubbed until later phases.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/simulations/scenarios": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_api_v1_simulations_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulations/runs": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Start Simulation",
        "operationId": "start_simulation_api_v1_simulations_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulations/runs/{run_id}": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Get Simulation Run",
        "operationId": "get_simulation_run_api_v1_simulations_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulations/runs/{run_id}/reset": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Reset Simulation Run",
        "operationId": "reset_simulation_run_api_v1_simulations_runs__run_id__reset_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulations/runs/{run_id}/faults": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Create Fault",
        "operationId": "create_fault_api_v1_simulations_runs__run_id__faults_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFaultRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulations/runs/{run_id}/faults/{fault_id}": {
      "patch": {
        "tags": [
          "simulation"
        ],
        "summary": "Patch Fault",
        "operationId": "patch_fault_api_v1_simulations_runs__run_id__faults__fault_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "fault_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Fault Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchFaultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFaultRequest": {
        "properties": {
          "fault_type": {
            "$ref": "#/components/schemas/FaultType"
          },
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fault_type",
          "outlet_id"
        ],
        "title": "CreateFaultRequest"
      },
      "CreateRunRequest": {
        "properties": {
          "scenario_code": {
            "$ref": "#/components/schemas/ScenarioCode"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "config_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Overrides"
          },
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id",
            "default": "0b000000-0000-0000-0000-000000000001"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_code"
        ],
        "title": "CreateRunRequest"
      },
...
      "FaultSummary": {
        "properties": {
          "fault_injection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Fault Injection Id"
          },
          "fault_type": {
            "$ref": "#/components/schemas/FaultType"
          },
          "outlet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Outlet Id"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fault_injection_id",
          "fault_type",
          "outlet_id",
          "parameters",
          "is_enabled",
          "scheduled_at"
        ],
        "title": "FaultSummary"
      },
      "FaultType": {
        "type": "string",
        "enum": [
          "delay",
          "missing_feed",
          "missing_field",
          "conflicting_balance",
          "malformed_payload"
        ],
        "title": "FaultType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchFaultRequest": {
        "properties": {
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "is_enabled"
        ],
        "title": "PatchFaultRequest"
      },
...
      "RunArtifactCounts": {
        "properties": {
          "ingestion_batches": {
            "type": "integer",
            "title": "Ingestion Batches",
            "default": 0
          },
          "ingestion_events": {
            "type": "integer",
            "title": "Ingestion Events",
            "default": 0
          },
          "transactions": {
            "type": "integer",
            "title": "Transactions",
            "default": 0
          },
          "cash_snapshots": {
            "type": "integer",
            "title": "Cash Snapshots",
            "default": 0
          },
          "provider_snapshots": {
            "type": "integer",
            "title": "Provider Snapshots",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunArtifactCounts"
      },
      "RunResponse": {
        "properties": {
          "simulation_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Simulation Run Id"
          },
          "scenario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scenario Id"
          },
          "scenario_code": {
            "$ref": "#/components/schemas/ScenarioCode"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "config_snapshot": {
            "type": "object",
            "title": "Config Snapshot"
          },
          "status": {
            "$ref": "#/components/schemas/SimulationRunStatus"
          },
          "started_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By User Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "faults": {
            "items": {
              "$ref": "#/components/schemas/FaultSummary"
            },
            "type": "array",
            "title": "Faults"
          },
          "artifacts": {
            "$ref": "#/components/schemas/RunArtifactCounts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "simulation_run_id",
          "scenario_id",
          "scenario_code",
          "seed",
          "config_snapshot",
          "status",
          "started_at"
        ],
        "title": "RunResponse"
      },
      "ScenarioCode": {
        "type": "string",
        "enum": [
          "normal",
          "scenario_a",
          "scenario_b",
          "scenario_c",
          "scenario_d"
        ],
        "title": "ScenarioCode"
      },
      "ScenarioListResponse": {
        "properties": {
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioResponse"
            },
            "type": "array",
            "title": "Scenarios"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenarios"
        ],
        "title": "ScenarioListResponse"
      },
      "ScenarioResponse": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scenario Id"
          },
          "code": {
            "$ref": "#/components/schemas/ScenarioCode"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "default_seed": {
            "type": "integer",
            "title": "Default Seed"
          },
          "default_config": {
            "type": "object",
            "title": "Default Config"
          },
          "validation_split": {
            "$ref": "#/components/schemas/ValidationSplit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id",
          "code",
          "name",
          "description",
          "default_seed",
          "default_config",
          "validation_split",
          "is_active"
        ],
        "title": "ScenarioResponse"
      },
...
      "SimulationRunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "reset"
        ],
        "title": "SimulationRunStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationSplit": {
        "type": "string",
        "enum": [
          "tuning",
          "held_out",
          "demo"
        ],
        "title": "ValidationSplit"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}